met crate tools reorg
