+++ framework/meta/src/tools/mod.rs
@@ -1,0 +1,4 @@
+mod git_describe;
+pub mod post_build;
+
+pub use git_describe::git_describe;
